prettier issue fixed.

+++ contracts/test/oracles/default/UniswapTwapPriceOracleV2Resolver.t.sol
@@ -36,19 +36,26 @@ contract UniswapTwapOracleV2ResolverTest is BaseTest {
     address STELLA_WGLMR = 0x7F5Ac0FC127bcf1eAf54E3cd01b00300a0861a62; // STELLA/WGLMR
     address CELR_WGLMR = 0xd47BeC28365a82C0C006f3afd617012B02b129D6; // CELR/WGLMR
     address wglmr = 0xAcc15dC74880C9944775448304B263D191c6077F; // WBNB
-    uint observationCount = twapPriceOracleRoot.observationCount(STELLA_WGLMR);
+    uint256 observationCount = twapPriceOracleRoot.observationCount(STELLA_WGLMR);
     emit log_named_uint("STELLA_WGLMR observationCount: ", observationCount);
     {
       (, , uint32 lastTime) = IUniswapV2Pair(STELLA_WGLMR).getReserves();
       emit log_named_uint("STELLA_WGLMR last time: ", lastTime);
-      (uint32 timestamp, uint256 price0Cumulative, uint256 price1Cumulative) = twapPriceOracleRoot.observations(STELLA_WGLMR, 0);
+      (uint32 timestamp, uint256 price0Cumulative, uint256 price1Cumulative) = twapPriceOracleRoot.observations(
+        STELLA_WGLMR,
+        0
+      );
       emit log_named_uint("STELLA_WGLMR observations timestamp: ", timestamp);
-      emit log_named_bytes("STELLA_WGLMR observations timestamp diff: ", abi.encode(block.timestamp - timestamp > 15 minutes));
+      emit log_named_bytes(
+        "STELLA_WGLMR observations timestamp diff: ",
+        abi.encode(block.timestamp - timestamp > 15 minutes)
+      );
       emit log_named_uint("STELLA_WGLMR observations price0Cumulative: ", price0Cumulative);
     }
-    UniswapTwapPriceOracleV2Resolver.PairConfig[] memory pairConfigs = new UniswapTwapPriceOracleV2Resolver.PairConfig[](0);
+    UniswapTwapPriceOracleV2Resolver.PairConfig[]
+      memory pairConfigs = new UniswapTwapPriceOracleV2Resolver.PairConfig[](0);
     resolver = new UniswapTwapPriceOracleV2Resolver(pairConfigs, twapPriceOracleRoot);
-  
+
     UniswapTwapPriceOracleV2Resolver.PairConfig memory pairConfig = UniswapTwapPriceOracleV2Resolver.PairConfig({
       pair: STELLA_WGLMR,
       baseToken: wglmr,
